Add Plug.ErrorHandler
Adds Plug.ErroHandler to be able to deal with errors in production.
Worth noting is the fact that if used with Plug.Debugger, it should go
after it (or else Debugger would not work).

`handle_error/2` needs to take in a `conn` and a map with the kind of
error (e.g. :throw), the reason for the error which could be an
exception or a different term, and the stacktrace.